requirements.txt
requests
dnspython

